+++ cli/src/commonMain/kotlin/com/kss/zoom/cli/ZoomCommand.kt
@@ -2,8 +2,12 @@ package com.kss.zoom.cli
 
 import com.github.ajalt.clikt.core.CliktCommand
 import com.github.ajalt.clikt.core.context
+import com.github.ajalt.clikt.core.subcommands
+import com.github.ajalt.mordant.terminal.Terminal
+import com.kss.zoom.Zoom
+import com.kss.zoom.cli.subcommands.*
 
-class ZoomCommand : CliktCommand(
+class ZoomCommand(zoom: Zoom, terminal: Terminal) : CliktCommand(
     help = """
         CLI for Zoom API
     """.trimIndent(),
@@ -16,8 +20,19 @@ class ZoomCommand : CliktCommand(
         context {
             helpOptionNames = emptySet()
         }
-//        completionOption()
+
+        subcommands(
+            HelpCommand(terminal),
+            VerboseCommand(),
+            ColorsCommand(),
+            ListMeetingsCommand(zoom),
+            LoginCommand(zoom, terminal)
+        )
     }
+
+    fun authCommands(): List<AuthCommand> =
+        registeredSubcommands().filterIsInstance<AuthCommand>()
+
     override fun run() {
     }
 }
